backend/requirements.txt
# Core pipeline
typer>=0.12
rich>=13.7
httpx>=0.27
pydantic>=2.7

# Agent layer
pydantic-ai-slim[openai]>=0.0.14
pypdf>=4.0
reportlab>=4.0
icalendar>=5.0
python-dotenv>=1.0
jinja2>=3.1

# External APIs (Google Places / Gmail)
googlemaps>=4.10
google-auth>=2.0
google-auth-oauthlib>=1.2
google-api-python-client>=2.100

# Web dashboard
streamlit>=1.30
plotly>=5.18
py3Dmol>=2.0

# Optional: ML scoring (uncomment to enable MHCflurry)
# mhcflurry>=2.1
